restruct
