+++ docbrown/core/src/tadjset.rs
@@ -222,6 +222,7 @@ impl<V: Ord + Into<usize> + From<usize> + Copy + Hash, Time: Copy + Ord> TAdjSet
         }
     }
 }
+
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
 pub(crate) struct Edge<V: Clone + PartialEq + Eq + PartialOrd + Ord> {
     v: V,
